Change code to support the profilers move
