+++ gprofiler/profilers/ruby.py
@@ -11,7 +11,7 @@ from psutil import Process
 from gprofiler.exceptions import ProcessStoppedException, StopEventSetException
 from gprofiler.log import get_logger_adapter
 from gprofiler.merge import parse_and_remove_one_collapsed
-from gprofiler.profiler_base import ProcessProfilerBase
+from gprofiler.profilers.profiler_base import ProcessProfilerBase
 from gprofiler.types import StackToSampleCount
 from gprofiler.utils import pgrep_maps, random_prefix, resource_path, run_process
